fix the parsing of the number of the chosen island.
It was wrong since it was taking the last character of the string, but the number of the island can have also 2 digit, so the used approach was wrong.
Now it is used the "#" to split the string into 2 parts (the part containing letters and the one containing numbers), and then it is parsed into integer the second part of the original string.